backend/requirements.txt
django
djangorestframework
mysqlclient
djangorestframework-simplejwt
drf-yasg
django-environ
